requirements.txt
asgiref==3.3.0
Django==3.1.3
django-crispy-forms==1.9.2
pytz==2020.4
sqlparse==0.4.1

djangorestframework~=3.12.4